+++ src/scripts/tidy5e-sheet.js
@@ -214,10 +214,23 @@ export class Tidy5eSheet extends ActorSheet5eCharacter {
 			const html = await super._renderInner(...args);
 			
 			try {
-				const actionsTab = html.find('.actions-target');
+				
+				// Update the nav menu
+				const actionsTabButton = $('<a class="item" data-tab="actions">' + game.i18n.localize(`DND5E.ActionPl`) + '</a>');
+				const tabs = html.find('.tabs[data-group="primary"]');
+				tabs.prepend(actionsTabButton);
+
+				// Create the tab
+				const sheetBody = html.find('.sheet-body');
+				const actionsTab = $(`<div class="tab actions" data-group="primary" data-tab="actions"></div>`);
+				const actionsLayout = $(`<div class="list-layout"></div>`);
+				actionsTab.append(actionsLayout);
+				sheetBody.prepend(actionsTab);
+
+				// const actionsTab = html.find('.actions-target');
 				
 				const actionsTabHtml = $(await CAL5E.renderActionsList(this.actor));
-				actionsTab.html(actionsTabHtml);
+				actionsLayout.html(actionsTabHtml);
 			} catch (e) {
 				// log(true, e);
 			}