+++ changelog.md
@@ -18,11 +18,10 @@ Thank you very much!
 
 ### Version 0.4.15 ###
 
-- integrated character actions list into the PC sheet
+- integrated actor actions list module
 - fix for Portrait Popout and CUB not hiding Actor names
 - new GM setting "total edit lock" to lock down sensitive fields when the edit lock is enabled
 - new Debug setting to show a marker for items with active effects (for @tposney)
-- added new tidy icons font
 
 ### Version 0.4.14 ###